0.2.18

+++ common/models/project.py
@@ -76,7 +76,7 @@ class Reading(FirebaseObject):
     script_id: Optional[str] = None
     type: ReadingType
     info: ReadingAvatarInfo
-    duration: Optional[int] = None
+    duration: Optional[float] = None
     status: ReadingStatus = ReadingStatus.new
     error: Optional[str] = None
     assets: Optional[List[Asset]] = None
